Changes to SelectDataPlugin
Overhaul of ResultTable
Reworking the SelectDataNodeModel to work with ResultTable
Fixed bug in which ResultTable wasn't filling in the table correctly.
Changes to SelectDataNodeDialog filling in list of nodetypes while connecting to jevis.
Fixed bug where Dialog Pane wouldn't load cause of not existing list.
Overhaul of the SearchforNodesClass to work with ResultTable.
Changes to SearchforStructure and SearchForNodes to get information from DataBaseConfiguration class.