Feature and Bugfixes
Feature:
Added a new filtering system in the Datapoint Search to only get useable locations, components and devices to select.

Bugfixes:
Fixed a bug where GetDataNode would also connect to DataBase when Default Connection information is empty.
Fixed a bug where not enough information was passed between the nodes to connect to Database.